Tout compile, mais ça marche pas (AleaNode n'accède pas à Mat, et les heuristiques sont pas bien transmises

+++ Sources/AleaNode.h
@@ -5,16 +5,7 @@
 #ifndef GHOSTBUSTER_ALEANODE_H
 #define GHOSTBUSTER_ALEANODE_H
 
-#include "sdlclg.h"
-#include "Constantes.h"
-#include "Pacman.h"
-#include "Personnage.h"
-#include "Fantome.h"
-#include "GhostNode.h"
-#include "PacNode.h"
 #include "Node.h"
-
-#include <vector>
 
 class AleaNode : public Node
 {
@@ -25,7 +16,7 @@ private:
     int profondeur; // La profondeur de l'arbre (décroissante, feuille -> 0)
     Node& parent; // Référence vers le parent, pour lui transmettre les infos
     Direction choix; // La direction choisie par le Pacman, à retransmettre au parent en cas de meilleure heuristique
-    int heuristique;
+    double  heuristique;
     int ghost; // Pour savoir à quel ghost du tableau on est (et quand s'arrêter)
     double proba; // Proba de cette branche
     double possib; // Nombre de possibilités pour le fantôme en cours
