+++ Sources/GhostNode.h
@@ -5,16 +5,7 @@
 #ifndef GHOSTBUSTER_GHOSTNODE_H
 #define GHOSTBUSTER_GHOSTNODE_H
 
-#include "sdlclg.h"
-#include "Constantes.h"
-#include "Pacman.h"
-#include "Personnage.h"
-#include "Fantome.h"
-#include "PacNode.h"
-#include "AleaNode.h"
 #include "Node.h"
-
-#include <vector>
 
 class GhostNode : public Node
 {
@@ -25,7 +16,7 @@ private:
     int profondeur; // La profondeur de l'arbre (décroissante, feuille -> 0)
     Node& parent; // Référence vers le parent, pour lui transmettre les infos
     Direction choix; // La direction choisie par le Pacman, à retransmettre au parent en cas de meilleure heuristique
-    int heuristique;
+    double  heuristique;
 public:
     GhostNode(const array<Fantome, 4>& f, const Pacman& p, array<array<Objet, Largeur>, Hauteur> n,
               int pr, Node& pa, Direction d);
